Fix compiler warning

+++ model/minsky.cc
@@ -641,7 +641,7 @@ namespace minsky
                              {
                                destTable.cell(row, col).clear();
                                // if this leaves an empty row, delete entire row
-                               for (int c=0; c<destTable.cols(); ++c)
+                               for (size_t c=0; c<destTable.cols(); ++c)
                                  if (!destTable.cell(row, col).empty())
                                    goto rowNotEmpty;
                                rowsToDelete.push_back(row);
